% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladedate.R
\name{clade.date}
\alias{clade.date}
\title{Empirical Calibration Density Generator}
\usage{
clade.date(
  ages,
  p = c(0, 0.5, 0.95),
  n = 10000,
  method = "StraussSadler",
  KStest = FALSE,
  PDFfitting = "best",
  ...
)
}
\arguments{
\item{ages}{vector of fossil ages (if ages are known exactly) or or a matrix with fossils in rows and two columns: the first with the minimum age bounds (upper stratigraphic bounds) and the second with the maximum age bounds (lower stratigraphic bounds).}

\item{p}{vector of probabilities indicating the quantiles to be reported.}

\item{n}{number of Monte Carlo replicates for generating the empirical distribution.}

\item{method}{a character string specifying the method: \code{"StraussSadler"} (default), \code{"RobsonWhitlock"}, \code{"Beta"}, \code{"NorrisPenGap"}, \code{"NorrisGhostLin"}, or \code{"OLE"}. See \link{pdate} for details.}

\item{KStest}{if TRUE, a Kolmogorov-Smirnov test is returned testing the null hypothesis that the distribution of fossil ages is uniform.}

\item{PDFfitting}{a character string specifying a standard probability density function ( \code{"lognormnal"}, \code{"gamma"}, \code{"exponential"}) to be fit to the sample. "best" returns the best-fit density among those. “skewStudent” fits a skew-Student distribution using function \code{sstdFit} in the \code{fGarch} library. NULL prevents fitting.}

\item{...}{additional arguments passed to \code{qdate} (see \link{pdate} for details)}

\item{plot}{If TRUE, plots fossil ages, a histogram of the Monte Carlo replicates, the median, and the 95\% quantile (and the fitted probability density function if available).}

\item{repvalues}{If TRUE, all n Monte Carlo replicates are returned.}

\item{breaks}{Specifies the algorithm used to determine the number of histogram cells in the plot (see \link{hist} for alternatives, if needed).}
}
\value{
a list with the following elements:
\itemize{
\item{Quantiles}
\item{KStest}
\item{PDFfit}
}
}
\description{
This function uses the fossil record of a clade to generate an empirical probability density function that can be used as a calibration density or calibration prior in Bayesian divergence time estimation. Relevant quantiles such as the median or the 95\% quantile can also obtained to be used in analyses based on fixed ages or bounds (e.g. penalized likelihood).
}
\details{
This function generates an empirical distribution representing the probability density of the age of a clade of organisms. The distribution is generated using a Monte Carlo algorithm based on alternative methods of estimating the age of a clade from a set of fossils ages (see \link{pdate}). The algorithm incorporates fossil age uncertainty (if present) by sampling ages from the time interval were the fossil was found.
}
\examples{

## Matrix of fossil ages including maximum and minimum bounds for each
Fages <- cbind(c(50, 38, 30, 14, 3.5), c(56, 45, 35, 14, 6))

## Basic execution
  clade.date(Fages)

## Alternative options
  clade.date(Fages, method="OLE", plot=TRUE, PDFfitting="lognormal")

}
\references{
Claramunt, S. & J. L. Cracraft. (2015) A new time tree reveals Earth history’s imprint on the evolution of modern birds. \emph{Science Advances} \strong{1}:e1501005 \url{https://doi.org/10.1126/sciadv.1501005}
}
\seealso{
\link{pdate} for point estimates when fossil ages are known exactly, quantile functions, and the random generator function.
}
\author{
Santiago Claramunt, \email{sclaramunt@rom.on.ca}
}
