% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil.record.R
\name{fossil.record}
\alias{fossil.record}
\title{Extract Fossil Record From Tree}
\usage{
fossil.record(x, phy, fossils)
}
\arguments{
\item{x}{the node number that defines the crown-clade}

\item{phy}{a phylogenetic tree of class 'phylo'}

\item{fossils}{an object of class 'fossils' produced by \link{FossiSim}}
}
\value{
a vector of fossil ages
}
\description{
Thus function collects the fossil record (ages) of a particular clade from the result simmulations in the package \link{FossilSim}
}
\details{
This function extract the fossil record of a clade from a \link{FossiSim} object. The clade is indicxated by the node that represent the most recent common ancestor of the clade.
}
