% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clade.date.R
\name{summary.clade.date}
\alias{summary.clade.date}
\title{Print Summary of an Empirical Calibration Density}
\usage{
summary.clade.date(x, param = "standard", digits = 4)
}
\arguments{
\item{x}{an object generated by \link{clade.date}.}

\item{param}{a character string specifying the format for the report of the estimated parameters matching that of common programs for Bayesian time-tree estimation: \code{"standard"} or \code{"mrbayes"}.}

\item{digits}{number of significant digits to report (default to 4)}
}
\value{
a list with the following elements:
\itemize{
\item{Quantiles}
\item{KStest}
\item{PDFfit}
}
}
\description{
This function prints a summary of an empirical calibration density generated by \code{"clade.date"}.
}
\details{
Parameters of the probability density function fit, if present, are reported their standard parameterizations (\code{"standard"} see \link{Distributions}) or the parameterization used in MrBayes (\code{"mrbayes"}), in which the gamma and lognormal distributions are parameterized by their mean and standard deviation.
}
\examples{

## Matrix of fossil ages including maximum and minimum bounds for each
Fages <- cbind(c(50, 38, 30, 14, 3.5), c(56, 45, 35, 14, 6))

## Basic execution
  clade.date(Fages)

## Alternative options
  X <- clade.date(Fages, method="RobertsSolow", plot=TRUE, PDFfitting="lognormal")
  summary(X)

}
\seealso{
\link{clade.date}.
}
\author{
Santiago Claramunt, \email{s.claramunt@utoronto.ca}
}
\references{
Claramunt, S. (2022) CladeDate: calibration information generator for divergence time estimation. \emph{Methods in Ecology and Evolution} \strong{13}, 2331-2338 \url{https://doi.org/10.1111/2041-210X.13977}
}
