% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clade.date.R
\name{plot.clade.date}
\alias{plot.clade.date}
\title{Plot Empirical Calibration Densities}
\usage{
plot.clade.date(object, breaks = "FD", ...)
}
\arguments{
\item{object}{an object of class \code{"clade.date"}.}

\item{breaks}{Specifies the algorithm used to determine the number of histogram cells in the plot (see \link{hist} for alternatives, if needed).}

\item{...}{additional arguments passed to \link{par}.}
}
\value{
Creates a plot from a \code{clade.date} object.
}
\description{
This function plots empirical calibration densities generated using \link{clade.date}.
}
\details{
This function plots empirical calibration densities from a \code{clade.date} object.
}
\examples{

## Matrix of fossil ages including maximum and minimum bounds for each
Fages <- cbind(c(50, 38, 30, 14, 3.5), c(56, 45, 35, 14, 6))

## Clade.date analysis
X <- clade.date(Fages)

## Plot
plot(X)

}
\references{
Claramunt, S. (2022) CladeDate: calibration information generator for divergence time estimation. \emph{Methods in Ecology and Evolution} \strong{13}, 2331-2338 \url{https://doi.org/10.1111/2041-210X.13977}}
\seealso{
\link{clade.date} for generating empirical calibration densities.
}
\author{
Santiago Claramunt, \email{s.claramunt@utoronto.ca}
}
